move from build a tarball to building from git
While building the tarball makes it more transparant
to apply patches since they must be listed in the
PKGBUILD, this approach misses a lot of code that
landed in the upstream stable branch intentionally.
Also it gives room for discussions what patches to
apply and when they're ready to be used. This avoids
confusion by setting the upstream merge of any patch
as the requirement to be applied. ymmv